+++ grpcclient/grpcclient.go
@@ -91,14 +91,19 @@ func (cfg *Config) CallOptions() []grpc.CallOption {
 	return opts
 }
 
-// DialOption returns the config as a grpc.DialOptions.
-func (cfg *Config) DialOption(unaryClientInterceptors []grpc.UnaryClientInterceptor, streamClientInterceptors []grpc.StreamClientInterceptor) ([]grpc.DialOption, error) {
-	var opts []grpc.DialOption
-	tlsOpts, err := cfg.TLS.GetGRPCDialOptions(cfg.TLSEnabled)
+// DialOption returns the config as a slice of grpc.DialOptions.
+//
+// keepaliveTime is the number of seconds after which the client will ping the server in case of inactivity.
+// See `google.golang.org/grpc/keepalive.ClientParameters.Time` for reference.
+//
+// keepaliveTimeout is the number of seconds the client waits after pinging the server, and if no activity is
+// seen after that, the connection is closed. See `google.golang.org/grpc/keepalive.ClientParameters.Timeout`
+// for reference.
+func (cfg *Config) DialOption(unaryClientInterceptors []grpc.UnaryClientInterceptor, streamClientInterceptors []grpc.StreamClientInterceptor, keepaliveTime, keepaliveTimeout int64) ([]grpc.DialOption, error) {
+	opts, err := cfg.TLS.GetGRPCDialOptions(cfg.TLSEnabled)
 	if err != nil {
 		return nil, err
 	}
-	opts = append(opts, tlsOpts...)
 
 	if cfg.BackoffOnRatelimits {
 		unaryClientInterceptors = append([]grpc.UnaryClientInterceptor{NewBackoffRetry(cfg.BackoffConfig)}, unaryClientInterceptors...)
@@ -140,13 +145,29 @@ func (cfg *Config) DialOption(unaryClientInterceptors []grpc.UnaryClientIntercep
 
 	return append(
 		opts,
-		grpc.WithDefaultCallOptions(cfg.CallOptions()...),
-		grpc.WithChainUnaryInterceptor(unaryClientInterceptors...),
-		grpc.WithChainStreamInterceptor(streamClientInterceptors...),
-		grpc.WithKeepaliveParams(keepalive.ClientParameters{
-			Time:                time.Second * 20,
-			Timeout:             time.Second * 10,
+		withDefaultCallOptions(cfg.CallOptions()...),
+		withUnaryInterceptor(middleware.ChainUnaryClient(unaryClientInterceptors...)),
+		withStreamInterceptor(middleware.ChainStreamClient(streamClientInterceptors...)),
+		withKeepaliveParams(keepalive.ClientParameters{
+			Time:                time.Duration(keepaliveTime) * time.Second,
+			Timeout:             time.Duration(keepaliveTimeout) * time.Second,
 			PermitWithoutStream: true,
 		}),
 	), nil
+}
+
+var withDefaultCallOptions = func(cos ...grpc.CallOption) grpc.DialOption {
+	return grpc.WithDefaultCallOptions(cos...)
+}
+
+var withUnaryInterceptor = func(f grpc.UnaryClientInterceptor) grpc.DialOption {
+	return grpc.WithUnaryInterceptor(f)
+}
+
+var withStreamInterceptor = func(f grpc.StreamClientInterceptor) grpc.DialOption {
+	return grpc.WithStreamInterceptor(f)
+}
+
+var withKeepaliveParams = func(kp keepalive.ClientParameters) grpc.DialOption {
+	return grpc.WithKeepaliveParams(kp)
 }